fix(inventory): enforce security and normalize inventory dto types
- add authentication and branch access validation in ProductsController
- remove duplicated @Version declarations from inventory JPA entities
- unify catalog handling in a single controller method without breaking route compatibility
- replace raw batch quantity with StockMovementQuantity and normalize money/date/id DTO types

+++ src/main/java/com/tuxlogic/shiftiq/platform/inventory/interfaces/rest/resources/ProductResource.java
@@ -1,6 +1,7 @@
 package com.tuxlogic.shiftiq.platform.inventory.interfaces.rest.resources;
 
 import java.util.UUID;
+import java.math.BigDecimal;
 
 public record ProductResource(
         UUID id,
@@ -9,7 +10,7 @@ public record ProductResource(
         String name,
         String sku,
         String description,
-        Double salePrice,
+        BigDecimal salePrice,
         Integer minimumStock,
         Integer currentStock,
         Boolean lowStockAlert
